+++ packages/yourloops/test/components/index.ts
@@ -29,10 +29,14 @@
 import testDatePickers from "./date-pickers";
 import testSnackbar from "./utils/snackbar.test";
 import TestPasswordStrengthMeter from "./password-strength-meter.test";
+import testSwitchRole from "./switch-role";
+import testDropdowns from "./dropdown";
 
 function testComponents(): void {
   describe("DatePickers", testDatePickers);
   describe("Snackbar", testSnackbar);
+  describe("Dropdowns", testDropdowns);
+  describe("SwitchRole", testSwitchRole);
   describe("Password strength meter", TestPasswordStrengthMeter);
 }