+++ CHANGELOG.md
@@ -5,12 +5,14 @@ It is based on Tidepool Blip 1.27.
 ## 2.0.6-beta4 - 2022-02-09
 ### Added
 - YLP-1105 Yourloops new blue theme
+- YLP-1106 Switch account from caregiver to hcp should ask for the profession
 - YLP-1122 Patient should enter the team unique ID to accept a team invitation
 - YLP-1188 Pink theme for clinical environment
 - YLP-1190 Advertise user about account deletion
 - YLP-1217 Patient should be better informed that he should not create an account through web UI
 ### Improvements
 - YLP-22 Print PDF content cleaned and improved (removing useless data)
+- YLP-288 Improve date selection on Trends view
 - YLP-1120 Overview tab cleanup
 ### Fixed
 - YLP-1101 Yourloops logo redirect not working for HCPs